Fix repeated countries and extract sharedfunctions
